requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
reportlab==4.0.4
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.24.3
setuptools==69.5.1
